Перенес логику отрисовки всех эвентов и сортировку в controllers/trip.js
